requirements.txt
Click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pytz==2019.3
python-dateutil==2.8.0
six==1.12.0
Werkzeug==2.3.3
pylint==2.4.4
Pillow==9.3
ultralytics==8.0.9
nest-asyncio==1.5.6
pyngrok==6.0.0
